l4sys: specify ex_regs flag usage constraints
Only one of the L4_THREAD_EX_REGS_ARM(64)_SET_EL_* flags might be used.
Currently, combining SET_EL0 and SET_EL1 will be rejected by the
kernel. Still, we define it as undefined behaviour in case we want to
introduce an operation in the future that would use this last,
unallocated bit combination.

Change-Id: I7ff6567c6b6ed43e5cc5cefa454808d9367f3c59