getaddrinfo.c: Avoid misleading indentation warning
When compiling getaddrinfo.c with clang the -Wmisleading-indentation
option will cause a warning due to the indentation lining up with the
previous statement in the if block above.

For gcc the warning is blinded by the commented line. See also:
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=107162

Move the comment behind the function call to make both compilers happy.

Change-Id: I378b8d83632a5c07bb6ca4621b013f7f7c9031ec