uclibc: Remove unused mkstemp64
The file with the definiton was built, but the defines to declare the
prototype were not set in our setup. So enabling more extensive warnings
caused a warning here. Remove the file from the build. If we ever need
it we should add it while also adding the right defines (see stdlib.h).

Change-Id: Iae3bee2562a0dfcec0df834e19f10dc8132a31f2