libgcc: use proper instructions on armv6+
If a binary is compiled for armv6 and beyond, it will most probably not
run on previous architectures anyway. Because ldrex/strex is defined by
armv6, use these instructions directly. For armv5 architectures the
kernel library page is still needed because atomic instructions require
kernel support on these ancient architectures.

Change-Id: Ib000b73d39c08367b6a0ebecbdd692f412444913