libgcc: do not use syscall page for dmb
On no-MMU systems there is no syscall page available. The only
functional difference is between armv5 and later architectures.
Starting with armv6 the CP15DMB control register is available. Even if
there is a dedicated instruction starting with armv7, the control
register is still available for backwards compatibility. Effectively
there is no greater benefit of using the syscall page abstraction
except to run armv5 binaries unmodified which is not a use case for
L4Re.

Change-Id: I72371dd80307db1a647789c1be4aac515ea21c28