ned: replace manual delete and reference-counting by Ref_ptr
It is considered bad style to explicitly use delete expressions when it
can be avoided. Being trained on not using delete helps avoiding using
it accidentally in problematic cases.

Also, the manual reference-counting was hard to follow. It was not
immediately obvious that each add_ref() call was matched by a single
remove_ref() call at run-time.

Change-Id: Ie44ee8d28bf04b1ff2619702e422d6e0a6601541