moe: Remove NARGSAVE stack adjustment on mips
The MIPS Calling Convention requires that stack space for four registers
must be reserved by the caller. However, the __start entry point in
uClibc expects argc to be at 0($sp), i.e. it does not adhere to the
MIPS Calling Convention, which makes sense because it is an entry point
not a regular function.

Change-Id: I118dc026b32635e7837a11919b81a6903227f5cf