x86: properly align stack pointer on entering init()
On x86/amd64, the return address of the 'call' instruction is pushed to
the stack. The called function considers this when adapting the stack
pointer. Do the alignment in crt0.S to achieve a 16-byte stack alignment
which is required for executing certain SSE instructions.

For amd64, the alignment was already -- add an explaining comment.

Change-Id: I121f0474ba2a4db583e2d51e68bcda32619baeb1