l4re-util: disable exceptions and RTTI
This library is part of the L4Re libc and may be linked into
applications that do not necessarily have a C++ runtime environment.
Some parts do rely on exceptions in their API, though. Move them
completely into the header so that the built shared library is still
free of exceptions and RTTI.

Change-Id: Ia7c0c2732d4d82cf571ca4148a78cbb28d76a76b