Revert "avoid hiding of non-private methods (Autosar A10-2-1)"
This reverts commit 3f1d901730dadc013f661b93d9b72e8ea30a9482.
This reverts change I617118d7d1bc19d49f10250090de1e0dcb9b6da9.

The hiding and hidden method differ only in the return type. The return
type of the base class’s method is more generic, hence, accidental usage
either does not compile (because the programmer expected the more
specific type of the hiding method) or behaves the same as if the hiding
method was called. Renaming one of the methods would unnecessarily
complicate the interface of the deriving class.

Change-Id: I3749521fda081c2e638fe4721f77753dbac8c692