l4re_vfs: Fix reading of empty files.
Previously it was not possible to read() an empty file using the
standard POSIX API. However, doing so is required e.g. when starting
runtimes like python which read configuration files that may be empty.
The problem comes down to attach() operations with size = 0 failing.
This fix introduces a special handling of read() operations for files
with size = 0, so that such calls are directly answered with return
code 0 (EOF).

Change-Id: Ic125e310d5af055d4ed031eb7dfa8b759ded1fb2
Signed-off-by: Viktor Reusch <viktor.reusch@barkhauseninstitut.org>
Signed-off-by: Till Miemietz <till.miemietz@barkhauseninstitut.org>