uclibc/atexit: Resolve multiple issues with dynamic atexit
Allocation from the statically allocated atexit function slots has
to be done under the __atexit_lock to prevent different threads from
selecting the same free slot. Thus move it into the __new_exitfn(),
which acquires the __atexit_lock already. This also has the advantage
that not only __cxa_atexit is now covered but also on_exit.

Furthermore, as the previous approach skipped calling __new_exitfn()
if a slot from the static exit function table was allocated, cleanup
was not enabled, i.e. `__exit_cleanup` was not set to `__exit_handler`,
and so on exit the registered atexit functions were not invoked.

Do not free the dynamic exit function table as the backend may have
already terminated.

Change-Id: I52dc8ebed8416db6af4460f9d903416192b207f3