[API] VMX: ABI change for nested virtualization support
This is the baseline ABI change that enables future support for VMX
nested virtualization. The vCPU state structure version for IA-32 and
AMD64 is bumped to indicate the kernel ABI change.

The VMX user information structure is extended by a nested revision
identifier. This field serves two purposes: (a) A non-zero value
indicates that the nested virtualization support is actually implemented
by the kernel. (b) The user space uses the non-zero nested revision to
derive the value of the IA32_VMX_BASIC MSR (i.e. the VMCS revision
identifier part) that should be exposed to the guest VM (contrary to the
IA32_VMX_BASIC MSR defined by the physical CPU that is used by the host).

Four additional software-defined natural-width fields are also added to
the software VMCS. The purpose of these fields is to pass the necessary
information between the kernel and user space to handle certain nested
VM exits (e.g. nested EPT violation).

Change-Id: Ic865600ef4ce2887f3c6b3b0d538203055800820