[API] VMX: Decouple software and hardware VMCS, add dirty bitmap
Improve the performance of the VMX support, especially targeting the
execution of multiple VMs in parallel and future support for nested
virtualization.

A new kernel object type that represents the hardware VMCS (as an
instance of a hardware vCPU context) is introduced. This enables the
user space to choose the policy of mapping vCPUs to hardware VMCSs.

* When no vCPU is running, the kernel does not allocate any hardware
  VMCS by itself, thus lowering the baseline memory consumption.

* In a memory-constrained environment, a single hardware VMCS can be
  shared by multiple vCPUs (managed by a single VMM task).

* In the typical case, the user space would use a 1:1 mapping of vCPUs
  to hardware VMCSs, thus enabling efficient context switching between
  vCPUs of separate VMs without the contention on a single hardware
  VMCS per physical CPU.

The mechanism for associating the hardware VMCS objects to vCPUs is
similar to the mechanism of associating the user tasks to vCPUs.

To avoid unnecessary copying of unmodified fields from the software VMCS
to the hardware VMCS on each VM entry, the extended vCPU state contains
a bitmap of dirty software VMCS fields now, indicating a value changed
by user space.

(Note: When multiple vCPUs share a common hardware VMCS, the user space
must ensure that the VMCS fields which differ between the vCPUs are
properly copied to the hardware VMCS by setting the dirty bitmap
appropriately. The kernel does not implement any failsafe mechanism for
that.)

The offsets in the offset table of the software VMCS are computed
relative to the area where the values are physically stored now (not
relative to the beginning of the whole software VMCS as before). This
enables to use the same numerical offsets both for the values and for
the dirty bitmap.

The software VMCS offset table has also been reordered and extended
with ranges of the individual groups of VMCS fields. This enables the
user space to easily check which reads/writes are out-of-bound and
should be ignored (instead of overwriting random fields in the software
VMCS).

This change requires a minor change of the kernel ABI and adding a
single function to the l4sys API, but the overall design of the ABI
stays the same. Therefore the vCPU state structure version for IA-32
and AMD64 is bumped to indicate the kernel ABI change.

Change-Id: I61cd855a7aff690ae3e7b0783625144eb3ee580d