riscv: Add support for Sstc extension
The current implementation of the Sstc implementation in QEMU does not
respect the henvcfg.STCE extension disable bit. Therefore, even though
the vstimecmp mechanism of the Sstc extension is not enabled, it still
triggers timer interrupts in the guest. Resulting in the guest getting
stuck during boot, endlessly handling timer interrupts.

Change-Id: I2a526b49ce6892c926f56a5d7dab3db867ca1d74