libloader: support utcb allocation on no-MMU systems
Because the kernel chooses the address of the UTCB memory on no-MMU
systems, it's location must be known before populating the AUX vector
on the child stack. This requires to allocate the task up-front. The
new alloc_prog() app-model function does exactly this for the
Remote_app_model. The Local_app_model does not need to do anything
because the current task is used which already has UTCB memory.

Change-Id: Ie8698134916d791504800c76b74a8caf17d7e7f3