riscv: Detect multi-letter extensions
So far bootstrap was only able to detect single-letter standard
extensions. Furthermore, the isa_ext bitfield in the KIP only provided
storage for 32 extension IDs.

To support multi-letter extensions, the ISA string parsing code in
bootstrap was reworked and the isa_ext bitfield in the KIP was expanded
to allow for up to 224 extension IDs (rather arbitrarily chosen):
  - IDs 0-25 are reserved for single-letter RISC-V ISA extensions.
  - IDs starting from 26 represent multi-letter extensions. The
    assignment does not follow a defined order, IDs are simply assigned
    incrementally when new extensions are added.

Change-Id: I355d4dfa755432cc8e1703ea3464b888a020b336