libpthread: support utcb allocation on no-MMU systems
Unfortunately the approach differs fundamentally between both system
types. On MMU based systems the user space chooses a suitable virtual
address and the kernel maps the ku_mem there. On no-MMU systems it's
the reverse: the kernel chooses the address and user space records the
fact in the region manager.

Change-Id: I3067e66a406bda469623530cf97cbe310f7355b4