[API] platform_control/ia32: Fix incorrect usage of cpu_call_many
In preparation for upcoming fix for Cpu_call::cpu_call_many(async=true),
which then will require also for asynchronous calls that the cpu lock is
not held.

The existing implementation had two incorrect cpu call usages, which
could result in deadlocks:

1. Platform_control::system_suspend() makes an asynchronous cpu call to
   the boot CPU, but passes state via the captured extra pointer and
   expects a return value, also via extra. Therefore the cpu call would
   have to be synchronous. Another problem is that the CPU lock is held
   during the cpu call.

2. suspend_ap_cpus() makes an asynchronous cpu call with the CPU lock
   held. Since the code is executed in a DRQ handler, it is not possible
   to release the CPU lock.

It was therefore necessary to rework the system suspend handling. The
system suspend handling is now executed directly by the calling thread.

This implies introducing a new restriction for the user-space API:
The Platform_control::system_suspend() operation must only be invoked by
a thread running on the boot CPU. Furthermore, user-space is responsible
that the thread is not migrated to a different CPU during the suspend,
otherwise the system might get stuck.

The restriction seems reasonable, as a system suspend affects the entire
system, so must only be accessible to an actor that has control over
the entire system.

See ticket #CD-1244.

Change-Id: Iaa7a044890678c3c817b0e287f982e2d99b344ef