libunwind_llvm: Maintain common version
Maintaining a version of libunwind_llvm for each major version of LLVM
appears unnecessary. libunwind_llvm did not change much over the last
major versions of LLVM and does not seem to be tightly coupled to a
specific LLVM version. So having a common version for all versions of
LLVM should be fine.

Change-Id: Id3c3e182bceaa8047c3f9314f2a6a91d4a54ba41