uclibc-minimal - Add missing dependency l4sys
libc/src/libc/stdlib/malloc-standard/malloc.h (from uclibc) included in
calloc.c in the same directory, includes l4/sys/consts.h, which is part
of the l4sys package and thus l4sys is a depedency of uclibc and
uclibc-minimal. The former already has it, this change adds it to the
latter.

Change-Id: I915836735011276e8aaff0f845505decde9200ae