uclibc/buildsys: temporary fix to symlinks in builddir
Change Ide791319cbe95e24bbdd1da252eae3049d61b84c introduced a symlink
that is later replaced with a real directory. cp -srf does not real well
with this case and refuses to overwrite the symlink. Clean the
destination directory before symlinking to avoid this case.

Since the change was only life for a day remove the workaround again
after a week.

Change-Id: Ia3f9a651b884322baf64e0a1f0b09e6fed99c4a9