uclibc: fix static PIE build
The build is broken upstream too when dynamic loader debugging is
enabled. To work around this, forcefully disable the debug features
when compiling reloc_static_pie.c.

Additionally, our minimal-variant is built without TLS. In this case,
the definition of _dl_load_base needs to be done in reloc_static_pie.c.
This was moved to libc-tls.c but obviously fails to link because the
definition will be missing if TLS is disabled.

Change-Id: I66422cfd4525bc81cbe786d321a5d15a17949668