libunwind_llvm: Disable debug output
On arm32 libunwind_llvm wants to print debug output to stderr, due to
there being special ARM Exception Handling ABI code. Printing however is
not supported by libc_minimal, which is used in sigma0. Until a better
fix is in place for this we disable the debug output in libunwind_llvm
completely.

To disable debug output in libunwind_llvm, we had to apply another patch
to contrib.

Change-Id: If39c02c4ce1ff526bdf0eb737189a31114da0a72