[API] Increase the software-defined VMCS field indexes
Both some simulators (namely Bochs and Simics) and some latest
real-world Intel CPUs already define VMCS fields whose base index
surpasses the base index of the current L4Re-specific software-defined
VMCS fields (although there are no collisions of the actual fields
yet).

To avoid failing on an assertion in debug builds and to limit the
likelihood of surprises even for production builds in the near future,
increase the indexes of the L4Re-specific software-defined VMCS fields.
This should give us a headroom of at least 20 base indexes according to
the current Intel specs, hopefully enough for several years to come.

This is a change of the ABI as indicated by the updated vCPU state
version number, but there is actually no change in functionality.

Change-Id: I04e0758f370166aed43385cc2522df2be58d0a4b