l4sys: arm64: use memory clobber for cache maintenance
The cache maintenance loop passed the memory location as
pointer-to-unsigned dummy parameter to the inline assembly. This is not
sufficient to force the compiler to write back all possibly affected
data. Only variables of unsigned type are actually affected. Values of
other types can be assumed to be unaffected by the compiler. Play safe,
and use a general memory clobber.

Change-Id: I85d74fd1b83ccc43e73a8b082997f17e61ce3f2c