l4sys: arm64: optimize cache flush functions
There is no need to issue a DSB before doing cache maintenance. Memory
accesses and cache maintenance instructions are ordered from the cores
perspective. OTOH, the DSB after the loop is important. It ensures that
the maintenance instructions have completed. This is important when
external observers are involved, which is typically the whole point of
these functions.

Change-Id: I088f2a62d2cbc7595e49012c01f5106fd8c158d4