ipc_types: move non-generic functions out of Gen_fpage
The information stored in a `Snd_fpage` and `Rcv_fpage` object are quite
different. Because of this, many member functions of `Gen_fpage` were
not applicable to both. This commit moves many of them to `Snd_fpage`.

This commit also adds the new member `Rcv_fpage::forward_mappings()`.
Formerly, `is_compound()` was used as it looks up the same bit. However,
this bit has a very different meaning for `Rcv_fpage` objects, therefore
the new function is introduced. The meaning of this bit for `Rcv_fpage`
was established in change I98a22eae7b612fdace49c69bb4df90410a6fadba.

This commit also removes an unused and not really generic constructor
from `Gen_fpage`.

Change-Id: If464341f92fda95d8bcc9ac498b9e0bce1b61f4e