libc-headers: Have deps for all installed headers
In case the header install is interruped (for example, by pressing
Ctrl-C in a wrong moment), not all headers are installed, however, the
all prerequisites are fulfilled for the following make call, never fully
installing all remaining headers. Now check that all targets are there
such that those are installed.

Change-Id: Iccf43b4e69e574f7e550eaabaea5684f4c5b2cc4