uclibc-ng: Upstream compile fixes for Linux, not relevant for L4Re
time64_helpers.h: error out when attempting to compile with TIME64 for
                  32bit targets on Linux versions prior to 5.1.0

Using 64bit time on 32bit targets was only introduced in Linux 5.1.0.
On versions prior to that, compiling uClibc-ng produces incorrect
headers that cause the `clock_nanosleep` syscall to receive incorrect
arguments.

This change was previously introduced in `features.h`, but was reverted
due to inclusion errors when compiling other projects.

Change-Id: I7a7237ad788a520e58ead8b9a0f1dc937556ce01
Signed-off-by: Nadav Tasher <tashernadav@gmail.com>
Signed-off-by: Marcus Haehnel <marcus.haehnel@kernkonzept.com>