[API] l4sys: move object code to libc
For regular applications this move is transparent. There are a couple
of low level projects that do not link with regular libc that might be
affected, though.

Remove the temporary declaration of --link-libc from the ldscripts
Makefile. Since the flag is now used by l4sys, there is no need for a
dummy declaration any more.

Requires-Change-Id: Ic03e8adba5eb00f76c86c869514f4fcc5808f57d
Requires-Change-Id: Ifa338d7b54cc48e16bb842de3e1be3ca55a3ac3b
Change-Id: I769cdc9187f20f44e451cb047f82210254d37daf