uclibc: add infrastructure to merge libs into libc
In the long run, libc should become self contained. That is, everything
of L4Re that is required to implement libc should be included into the
library. On the other hand, we want to keep the separation of the
components in the source code of l4re-core.

As a first step, a custom '--link-libc' bid-bender flag is defined,
which is set when linking libc. PC files can use this flag to supply
object files or libraries for the link. For static libraries this
includes searching for them manually, though. The linker is not invoked
for them but the archiver which does not search the library paths.

Because bid-bender will complain if a command line option is used that
is unknown, define a dummy instance in ldscripts. In the next commit
where libl4sys starts using it, this temporary workaround is removed.

The code object is moved into libc.a/libc.so for regular applications.
Foundational applications like moe and sigma0 that are based on
libc_minimal find this code in libc_minimal_l4re because libc_minimal
does not have any L4Re specific code. This separation needs to be kept
because bootstrap requires a plain, OS agnostic libc.

Change-Id: I8b998fb9df36daafe1293013602d20d594be250d