uclibc: extend x86 FPU state
Apparently, the signal frame layout of Linux on x86 was created before
the fxsave instruction was available. The structure only allocates
space for the x87 FPU. As we will have to save the MMX, XMM and MXCSR
state too, the structure needs to be extended to the 32 bit fxsave
format.

In the long run, we should implement or own L4Re version of the signal
frame structures as we are not bound to the Linux implementation in any
way.

Change-Id: Ia68348807a90a846ca0c7f16766d9adb6e0cf1c0