l4re: add itas protocol and interface
Some POSIX APIs need support from ITAS towards libc. In particular, the
signal handling functions need to be implemented in ITAS and the public
POSIX APIs need to forward the requests.

This protocol is considered an implementation detail of our libc. It is
not intended to be used by any application directly.

Change-Id: Icf433b4ffb9ca7e9be7c03cdfc67481e890d5d74