libunwind_llvm: Make exceptions work even if relocated to lower address
pinfo->dlpi_addr is effectively _dl_load_base, i.e. a relative offset.
Depending on whether the binary is relocated to an address above
load_info.start (in libloader) or below, exception handling either works
or ends prematurely, respectively.

In the context of the contrib code, the comparison seems like an
optimization. Furthemore the assumption there seems to be that
pinfo->dlpi_addr is already an address and that PIE binaries and shared
objects are always linked to address 0.

Change-Id: I37e8d7318296b6c1e57d9bba812a4fe9a72a5229