l4re_itas: add proper server loop and registry
So far, the dispatch loop would only support one interface: the region
manager. But for POSIX signals we need to distinguish the different
threads and need to implement timers. For that, a full fledged server
loop is required.

For the registry we cannot use the L4Re::Util::Object_registry because
the ITAS has its own capability allocator. Therefore we have to roll
our own, minimal implementation.

Change-Id: I28854589091c0f85aa3bd098ccecf5e46d6b37f2