l4re_itas: implement POSIX signals
To properly attribute thread faults to the right thread cap, a distinct
IPC gate is used for each thread. The new signal manager implements the
L4Re::Itas interface so that it knowns all threads.

Besides synchronous signals, also asynchronous ones are supported. As a
first step, SIGALRM is implemented. In principle, the infrastructure
should support any other signal too, including queued RT signals.

To jump to the signal handler, a small assembler stub is used. This is
mainly there to save the floating point state. So far it is not
implemented for all architectures.

Requires-Change-Id: I19754d1c9eb79cc09ef60a5fcb1b2eea11ce6e9d
Change-Id: I0181afd18c50b70abe59553eaa16bd88df252a38