uclibc: fix global constructors in libc.so
L4Re uses global constructors in libc because we link some C++ based
libraries. If the compiler emits the calls to those constructors in the
.ctor section, we have to call them ourself when libc.so is
initialized. In case the compiler utilizes .init_array, the dynamic
linker will have already called them for us.

Upstream uclibc does not need this because it has no global
constructors.

Change-Id: I1da19e452e151917a62b020a56edbd6a8fcb6bf1