ldso: correctly claim caps range
The ldso needs to allocate capability slots when loading the first
libraries. Later, once libc is initialized, this is replaced by the
L4Re::Util::cap_alloc allocator.

So far, the hard coded caps range was intersecting with the one that
l4re_itas claims. To make matters worse, the first_free_cap in the
L4Re::Env was not adjusted either. This could lead to the allocators
colliding with each other. Fix it by using the first_free_cap() as
starting point for the allocator and by adjusting first_free_cap()
accordingly.

Change-Id: I4d23bd40dab39c28f606b89bce9b8d8d730ddd70