Basic select() implementation
Client-side select() that waits passively on an I/O operation/condition
readiness of sets of file descriptors. This is a baseline implementation
which is expected not to scale ideally with the number of opened file
descriptors (since the backends currently always signal readiness
regardless of which file descriptors are currently examined), but it is
sufficient for the current practical use cases. Future optimizations are
possible by trading off with the simplicity of the interface and memory
consumption.

The signaling part piggy-backs on the SIGURG signal of the almost
universally deprecated, but still de facto standardized System V user
space signaling. This enables the backends to simply include the
standard signal.h header and it is minimally intrusive. The signaling
is triggered by

  gsignal(SIGURG)

Change-Id: I327cc13af31d1035b3a1c9e9d1c995113c3d0a29