sys: Add loop_dbg() function to L4::Server
This function is like loop, but in case of exceptions, prints the
error message instead of silently discarding it. The printing is done
via a parameter that implements a printf() member function. E.g., a
`L4Re::Util::Dbg` can be used.

Also add an according wrapper to L4Re::Util::Registry_server.

Change-Id: Idbd7cd9fe5be1efe1005ea951901bec1d9efefae