l4re: make L4Re::Util::cap_alloc compatible to L4Re::Cap_alloc
There is no good reason why the generic capability allocator interface
is not implemented by the L4Re::Util default cap allocator. In fact,
relying on L4Re::Cap_alloc::get_cap_alloc() in L4Re interfaces does not
work because it requires a working C++ runtime which is not available
in libc.

Change-Id: I1a4da0114835e6c31ee85f0e4cd2d78167504c7d