libsupc++: provide __cxa_thread_atexit
If a thread_local object has a destructor, the compiler will emit calls
to __cxa_thread_atexit() to destroy them on thread termination. There
are different options where this symbol can come from. In our case,
because it is not provided by uclibc, we need to use the compatibility
layer in libsupc++ that utilizes pthread keys.

Unfortunately this requires some tweaking of the libstdc++
configuration. The config header was generated on Linux where glibc
provides a __cxa_thread_atexit_impl symbol. This is not available for
us and the respective define must be removed for L4Re.

Change-Id: I296d996e7e65c0829280e6e3e156ba84cb737f48