libpthread: fix TLS alignment on amd64 and x86
The Intel platforms use the TLS_TCB_AT_TP TLS data model. There, the
TLS register points to the TCB (thread control block). The TLS data of
the executable and initially loaded shared objects is located *before*
that pointer:

                               %gs (x86) / %fs (amd64)
                                         |
                                         v
   +----+----------+----------+----------+-----+
   | .. | mod2 TLS | mod1 TLS | main TLS | TCB |
   +----+----------+----------+----------+-----+

Now, when allocating the static TLS data, the alignment of the thread
TLS pointer needs to be preserved and satisfy the minimum of the TLS
section. The size of the TCB must thus be added *after* the size of the
TLS area has been calculated, according to the required alignment. The
reason is _dl_allocate_tls_storage() which calculates the address of
the TLS pointer as:

  (char *) result + size - TLS_TCB_SIZE;

and the result must still satisfy the TLS alignment.

The _dl_tls_static_size global variable stores some initial headroom.
In case libdl is not linked, this should probably be
TLS_STATIC_SURPLUS. If libdl is linked, this is initialized to a much
bigger value by replacing the variable with a different definition in
libdl.

Change-Id: Ia5b58912881d9228eca701ec8767285fff417754