libunwind_llvm/uclibc-minimal: provide locking stubs without threading
Provide stubs for pthread_rwlock_{rdlock,wrlock,unlock}() if no
threading is available (e.g. Moe, ITAS).

In libunwind_llvm, don't define LIBUNWIND_USE_WEAK_PTHREAD because this
uses the availability of the weak symbol 'pthread_create' to detect if
threading is used or not. But this does not work with static linking
because pthread_create() might be linked to the application while
pthread_rwlock_{rdlock,wrlock,unlock}() are missing.

The more correct solution will be to provide a libunwind_llvm variant
without threading.

See ticket #CD-1428.

Change-Id: Ib32b71c13f7a30b38b865f29f48e5c075753591b