l4sys: avoid implementation-defined behavior shifting negative values
Shifting negative values, whether to the left or to the right, depends
on the compiler implementation. Consider this when extracting the label
of an l4_msgtag_t.

There is no problem with setting the label as the left-shift operation
is performed on an unsigned integer type. Unfortunately the same trick
cannot be used for right shifting.

This change is not intended to change any functionality. Known gcc and
Clang versions do the correct thing without this change. If this code
is compiled with a compiler according to the C++20 standard, this change
is also not required.

Note that C23 still does not make any guarantee about bitwise shifting
of negative values.

Change-Id: I482d91e85980caba19af1daac5724ca77803b25f