uclibc: Track source links and relink if target disappears
In particular during checkbuild, if a symlink disappeared because its
target was removed the buildsystem would not detect this and fail to
find the source file. It would not recover from that.

Add a dependency to the sources of all symlinks to force the buildsystem
to rebuild the symlinks in case of vanishing files.

Change-Id: Ice5aac01df43e70fdacd7564ec910449d5bda052