uclibc-ng/ldso.h: factor out _dl_auxvt from ldso.h
The _dl_auxvt is also used on non shared builds to access the values of
the vector from getauxval. Putting the definitions in ldso.h leads to
issues because that header also implements fcntl.h functionality making
it incompatible with files that include that header directly.

Factor out the _dl_auxvt related declarations to a new header dl-auxvt.h
and include from ldso.h as well as include it directly from
__uClibc_main.c to make the definitions available.

Change-Id: Iad12a5d2c766f4cf16c560c592991f281fe8f6f4
Signed-off-by: Marcus Haehnel <marcus.haehnel@kernkonzept.com>