ldscripts: link PIE binaries to address 0
This is an established standard. Both GNU binutils and LLVM lld use a
base address of zero for PIE binaries. Worse, other parts of the
ecosystem have started to rely on this property by implicitly assuming
that __ehdr_start is linked to address 0.

The lld linker practically mandates the use of __ehdr_start by libc
because it does not follow the informal behaviour of GNU ld to put the
link time address of the dynamic section into the first GOT entry.
Instead, they favour the use of __ehdr_start to compute the relocation
at runtime.

Interestingly, this is also a prerequisite for the GNU ld to mark the
binary as ET_DYN. In case the link base address is not zero, GNU ld
will emit ET_EXEC binaries, even if the -pie switch was passed.

Change-Id: I4e93b4f82757f4f798462e41c38a65c6b69337ae