uclibc: use __ehdr_start to compute load address
So far, uclibc relied on the informal behaviour of GNU ld to put the
link address of the _DYNAMIC symbol in the first GOT entry. This does
not work with LLVM lld which does not follow this convention.
Consequently, glibc has abandoned its usage in favour of using
__ehdr_start as magic symbol to infer the load address [1].

Note that this implies that the link time address of __ehdr_start is
*always* 0! So far, this seems to be the case on all platforms.

See ticket #CD-1395.

[1] https://sourceware.org/bugzilla/show_bug.cgi?id=28203

Change-Id: Ibde1c8afd242af112f086636b97ba21f8bcfd895
Requires-Change-Id: I17d5dfffc3d362db6a6feb853cd690e3ebf1b62d
Requires-Change-Id: I825eb46aeced6e5043b9d53fa66ef7b93f29f0a2