Revert "uclibc: fix static-pie relocation on arm64"
Even though the upstream comment was confusing, the code itself was at
least partially correct. We "just" need to rely on a symbol that has a
link address of 0. Apparently, this is the case with _start for most
systems.

This reverts change I857e7e7fa9d4e703ef1caaee270d6c2edd6d3935.

Change-Id: I19ff744957974c56988d751ceeb3428792fe72ef