ldscripts: apply default address if lld image-base is not used
Having a default base address in the linker script is a bit tricky for
lld. The trick is to apply the default address only if the current
address is not zero (an image-base was passed to the linker). Otherwise
the current link address is not touched because it is assumed to be set
from the outside.

Change-Id: I048b0da30708ddbda061d3005831af346b0f8b21