uclibc-ng: Use __ehdr_start to compute load address
So far, uclibc relied on the informal behaviour of GNU ld to put the
link address of the _DYNAMIC symbol in the first GOT entry. This does
not work with LLVM lld which does not follow this convention.
Consequently, glibc has abandoned its usage in favour of using
__ehdr_start as magic symbol to infer the load address [1].

Note that this implies that the link time address of __ehdr_start is
*always* 0! So far, this seems to be the case on all platforms.

[1] https://sourceware.org/bugzilla/show_bug.cgi?id=28203

Change-Id: Ie9b93d6cccc9bf18d5b6b1d31f27968f951322a4
Signed-off-by: Marcus Haehnel <marcus.haehnel@kernkonzept.com>