uclibc-ng: stdlib.h: only expose reallocarray prototype when available
The reallocarray() primitive is only provided by the malloc-standard
implementation: neither malloc nor malloc-simple provide it. This
causes issues when building for example util-linux on noMMU platforms
as:

- noMMU platforms can't use malloc-standard, so either malloc or
  malloc-simple are used, which means reallocarray() is not
  implemented

- util-linux detects the absence of reallocarray(), and provides its
  own implementation, but the prototype clashes with the prototype in
  uClibc's <stdlib.h>

The combination of which causes the following build failure:

 static declaration of ‘reallocarray’ follows non-static declaration

To fix this, let's not expose the prototype of reallocarray() if we
don't provide the implementation for it.

Change-Id: I1bc477f05b3602feb4b93221eb107d301f43f9db
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Signed-off-by: Marcus Haehnel <marcus.haehnel@kernkonzept.com>