introduce Cap_base::invalid_cap_error() and use in L4Re::chkcap()
If a capability has the L4_INVALID_CAP_BIT set, the capability selector
may carry an error code. The function L4Re::chkcap() returns that code
in case the passed 'err' parameter is 0.

The L4_INVALID_CAP_BIT should be masked out from the error code stored
in the capability selector, otherwise the value of that error does not
make sense.

Usually the error code from an invalid capability selector is not used.

Change-Id: I8744d3395f5cd623ca9d49573116e3b5cbbf8672