l4re_itas, Ned: use more appropriate error code for open_file()
If chkcap() failed this usually means that the corresponding file was
not found in the namespace. Use -L4_ENOENT instead of 0 (which is
interpreted as -L4_ENOMEM).

Trying to start an application from Ned Without this change:

  server  | terminate called from 0x7000e0e6 after throwing an instance
            of 'N2L413Runtime_errorE'
  server  |   what: bad, unknown runtime error: rom/ex_tg-sever-1

Trying to start an application from Ned Without this change:

  server  | terminate called from 0x7000e0e6 after throwing an instance
            of 'N2L417Element_not_foundE'
  server  |   what: No such object found: rom/ex_tg-sever-1

The open_file() function in Moe already uses -L4_ENOENT.

Change-Id: I9d190462e4844c26b98cec4b7ac3a1ca84be09f7