Adapt musl libc to L4Re
This changes some of the L4Re internals, such as ITAS to be more
flexible and work with both C library implementations.

It also adapts some musl internals that were too Linux specific to
easily adapt to from the outside.

The libc_backends adaptations are required because musl uses
preprocessor defines to "implement" xx64 functions, leading to aliasing
and thus double definitions if left unguarded.

Change-Id: Ie8a368eecd6e912c91dd7938de216a65529dbb4a
Co-authored-by: Georg Kotheimer <georg.kotheimer@kernkonzept.com>
Signed-off-by: Marcus Haehnel <marcus.haehnel@kernkonzept.com>