l4sys/compiler.h: Add L4 specific DECLS defines
The defines for __BEGIN_DECLS and __END_DECLS are gnu specific
extensions and not portable. They should not be used in portable code
since some C libraries do not provide their implementation. For L4
specific code we provide the L4_BEGIN_DECLS and L4_END_DECLS
alternatives which can be used when including l4/sys/compiler.h.

We still also provide the __BEGIN_DECLS functions in that header for
backwards compatibility where it makes sense, but their use is strongly
discouraged.

Change-Id: Id957f6325d81e225bd3e00bb4d2a784d19cbb854